test: Load transaction chainId as uint64 (#1570)
from_json_tx_common parsed the transaction "chainId" with
from_json<uint8_t>, throwing `from_json<uint8_t>: value > 0xFF` for any
chain ID above 255. This crashed `evmone t8n` on common networks such as
Sepolia (11155111) and Arbitrum One (42161): the transaction is parsed
before t8n overrides chain_id from its args, so the parse failed first
and no output was produced.

Parse it as uint64_t, matching the Transaction::chain_id field type.

Fixes #1569.

+++ test/utils/statetest_loader.cpp
@@ -375,7 +375,7 @@ static void from_json_tx_common(const json::json& j, state::Transaction& o)
     o.nonce = from_json<uint64_t>(j.at("nonce"));
 
     if (const auto chain_id_it = j.find("chainId"); chain_id_it != j.end())
-        o.chain_id = from_json<uint8_t>(*chain_id_it);
+        o.chain_id = from_json<uint64_t>(*chain_id_it);
     else
         o.chain_id = 1;
